code enhancement
